Speicher den ganzen Mist

+++ src/Gui/Fliesen/FliesenModel.java
@@ -2,7 +2,7 @@ package Gui.Fliesen;
 
 public class FliesenModel {
     private int _preis = 0;
-    public void gesamtpreisBerechnen (int[] preise) {
+    public void gesamtpreisBerechnen (double[] preise) {
         for(int i = 0; i<preise.length; i++) {
             this._preis += preise[i];
         }
